pyproject.toml
[tool.poetry]
name = "diet-assist"
version = "0.1.0"
description = ""
authors = ["Sagi Ezri <sagi.ezri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
pillow = "^10.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
transformers = "^4.44.2"
openai = "0.28"
timm = "^1.0.9"
gtts = "^2.5.3"
pyttsx3 = "^2.91"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
